fix: move date-fns locales to shared
