+++ shared/i18n/locales/en_US/translation.json
@@ -189,10 +189,6 @@
   "{{ count }} member": "{{ count }} member",
   "{{ count }} member_plural": "{{ count }} members",
   "Group members": "Group members",
-  "Online now": "Online now",
-  "Online": "Online",
-  "{{ username }} created": "{{ username }} created",
-  "{{ username }} updated": "{{ username }} updated",
   "Icon": "Icon",
   "Show menu": "Show menu",
   "Choose icon": "Choose icon",
@@ -903,9 +899,13 @@
   "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.": "Webhooks can be used to notify your application when events happen in {{appName}}. Events are sent as a https request with a JSON payload in near real-time.",
   "Inactive": "Inactive",
   "Create a webhook": "Create a webhook",
-  "Online now • Viewed {{ lastViewedAt }} ago": "Online now • Viewed {{ lastViewedAt }} ago",
+  "Online now • Viewed {{ lastViewedAt }}": "Online now • Viewed {{ lastViewedAt }}",
   "Online now • Never viewed": "Online now • Never viewed",
-  "Online {{ lastActiveAt }} ago • Viewed {{ lastViewedAt }} ago": "Online {{ lastActiveAt }} ago • Viewed {{ lastViewedAt }} ago",
-  "Online {{ lastActiveAt }} ago • Never viewed": "Online {{ lastActiveAt }} ago • Never viewed",
+  "Online {{ lastActiveAt }} • Viewed {{ lastViewedAt }}": "Online {{ lastActiveAt }} • Viewed {{ lastViewedAt }}",
+  "Online {{ lastActiveAt }} • Never viewed": "Online {{ lastActiveAt }} • Never viewed",
+  "You created {{ createdAt }}": "You created {{ createdAt }}",
+  "{{ username }} created {{ createdAt }}": "{{ username }} created {{ createdAt }}",
+  "You updated {{ updatedAt }}": "You updated {{ updatedAt }}",
+  "{{ username }} updated {{ updatedAt }}": "{{ username }} updated {{ updatedAt }}",
   "Uploading": "Uploading"
 }
